# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vectorizer API",
    "description": "High-performance vector database engine with semantic search capabilities.\nSupports multiple embedding providers (TF-IDF, BM25, BERT, MiniLM) and\nprovides real-time indexing with HNSW optimization.\n",
    "version": "0.3.1",
    "contact": {
      "name": "Vectorizer Team",
      "url": "https://github.com/hivellm/vectorizer"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:15001/api/v1",
      "description": "Local development server"
    },
    {
      "url": "http://127.0.0.1:15001/api/v1",
      "description": "Local server alternative"
    }
  ],
  "paths": {
...
    "/collections": {
      "get": {
        "summary": "List Collections",
        "description": "Get all available collections",
        "operationId": "listCollections",
        "tags": [
          "Collections"
        ],
        "responses": {
          "200": {
            "description": "List of collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCollectionsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Collection",
        "description": "Create a new vector collection",
        "operationId": "createCollection",
        "tags": [
          "Collections"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_name}": {
      "get": {
        "summary": "Get Collection Info",
        "description": "Get detailed information about a specific collection",
        "operationId": "getCollection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionInfo"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Collection",
        "description": "Delete a collection and all its vectors",
        "operationId": "deleteCollection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteCollectionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_name}/vectors": {
      "get": {
        "summary": "List Vectors",
        "description": "List vectors in a collection",
        "operationId": "listVectors",
        "tags": [
          "Vectors"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of vectors to return",
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of vectors to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vectors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVectorsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Insert Texts",
        "description": "Insert text content with automatic embedding generation",
        "operationId": "insertTexts",
        "tags": [
          "Vectors"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsertTextsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Texts inserted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsertTextsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_name}/vectors/{vector_id}": {
      "get": {
        "summary": "Get Vector",
        "description": "Get a specific vector by ID",
        "operationId": "getVector",
        "tags": [
          "Vectors"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vector_id",
            "in": "path",
            "required": true,
            "description": "ID of the vector",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vector information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VectorInfo"
                }
              }
            }
          },
          "404": {
            "description": "Vector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Vector",
        "description": "Delete a specific vector",
        "operationId": "deleteVector",
        "tags": [
          "Vectors"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vector_id",
            "in": "path",
            "required": true,
            "description": "ID of the vector to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vector deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteVectorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Vector not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_name}/search": {
      "post": {
        "summary": "Search Vectors",
        "description": "Search for similar vectors using vector or text query",
        "operationId": "searchVectors",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchUnifiedRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/collections/{collection_name}/search/text": {
      "post": {
        "summary": "Search by Text",
        "description": "Search using text query (automatically embedded)",
        "operationId": "searchVectorsByText",
        "tags": [
          "Search"
        ],
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "description": "Name of the collection to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/intelligent_search": {
      "post": {
        "summary": "Intelligent Search",
        "description": "Advanced semantic search with AI-powered query expansion, intelligent deduplication, and Maximal Marginal Relevance (MMR) diversification. Automatically generates multiple query variations, applies domain-specific knowledge, and ensures diverse, high-quality results. Ideal for complex research queries and comprehensive information retrieval.",
        "operationId": "intelligentSearch",
        "tags": [
          "Intelligent Search"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntelligentSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intelligent search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntelligentSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCollectionRequest": {
        "type": "object",
        "required": [
          "name",
          "dimension",
          "metric"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Collection name",
            "pattern": "^[a-zA-Z0-9_-]+$"
          },
          "dimension": {
            "type": "integer",
            "minimum": 1,
            "maximum": 4096,
            "description": "Vector dimension"
          },
          "metric": {
            "$ref": "#/components/schemas/DistanceMetric"
          },
          "hnsw_config": {
            "$ref": "#/components/schemas/HnswConfig"
          }
        }
      },
      "CreateCollectionResponse": {
        "type": "object",
        "required": [
          "message",
          "collection"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message"
          },
          "collection": {
            "type": "string",
            "description": "Created collection name"
          }
        }
      },
...
      "ListCollectionsResponse": {
        "type": "object",
        "required": [
          "collections"
        ],
        "properties": {
          "collections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CollectionInfo"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          },
        "timestamp": {
          "type": "string",
          "format": "date-time",
          "description": "Error timestamp"
        }
      }
    },
...
  },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication"
      }
    }
  },
...
}